tspow fixes in itests. cleaned code for merge
